AR-618 moved editslide
