+++ src/components/slide/edit-slide.jsx
