+++ src/components/slide/edit-slide.spec.js
